requirements.txt
# Core ML dependencies
ultralytics
torch
torchvision
opencv-python
numpy
pandas
matplotlib
seaborn
pillow
scikit-learn
pyyaml
tqdm

# Optional training enhancements
albumentations
tensorboard

# Web server (optional, for --serve mode)
fastapi
uvicorn
python-multipart

